Add custom interaction matching.
Interactions (recorded request/response pairs) were only matched
according to request method and URL during replay. This limits
the usefulness of go-vcr for testing HTTP requests because the
interesting state is often in the headers or body. For example, a
idiomatic PUT action in a REST API may have the same URL but respond
differently based on the body.

This commit matches requests to interactions using a Matcher function.
Users of the go-vcr library can define and use custom matchers in
their tests by setting the matcher on the recorder with SetMatcher.