Drop dependency on etcd
We used to have an example with `etcd` interactions, but that
also pulls `etcd` as a dependency, which may not be desired if
we want to keep this library small and tight.

Remove the `etcd` example and retain it only as a reference in the
README if anyone wants to use that anyways.

Fixes #28