Save cassettes without any interactions
This is often required when writing table-driven tests where some of the test cases
don't produce any interactions (like when validation fails) but the setup and test
code is the same for every test case. In this case, on a replay these tests then
fail because the cassette for the test cannot be found.