on error, report which cassette was not found
Before:

  requested cassette not found

After:

  requested cassette not found: FILENAME

There is a caveat: this change tries to be as backwards compatible as possible,
but, as the change in recorder_test.go shows, it breaks direct (!=) comparison
with the sentinel error. On the other hand, one could argue that sentinel
errors should always be compared with errors.Is().

In case you are not comfortable accepting this PR, no problems!

Thanks for go-vcr, everybody I show it to starts using it :-)