Add option to matcher to ignore headers
The new matcher, which is more strict now, is breaking tests where some
of the headers are skipped (like secrets) or change on each request (like tracing
spans). The cassete already supports the User-Agentt, and the Authorization has
also been added but not released yet.

In order to make it more generic and usable for people, I've added a more
generic implementation of these options, which will allow us to define which
headers we really want to ignore.

Signed-off-by: ivangonzalezacuna <ivangonzalezacuna@gmail.com>